# Conflicts:
#	.vs/GOproject/v17/.suo
#	.vs/GOproject/v17/fileList.bin
#	Assembly-CSharp-Editor.csproj
#	Assembly-CSharp.csproj
#	Library/ArtifactDB
#	Library/Artifacts/1a/1a78e30f6784c28b4d9ebf00cb5c4d79
#	Library/Artifacts/4d/4d9308db9cce81dbd0274224e5fde079
#	Library/Artifacts/5f/5ffc14354026f3c3f5862bc44b53f62c
#	Library/Artifacts/67/67c3cc6e974dc34f362992acfef7f27d
#	Library/Bee/1900b0aE-inputdata.json
#	Library/Bee/1900b0aE.dag
#	Library/Bee/1900b0aE.dag.json
#	Library/Bee/1900b0aE.dag.outputdata
#	Library/Bee/1900b0aE.dag.payloads
#	Library/Bee/1900b0aE.dag_derived
#	Library/Bee/TundraBuildState.state
#	Library/Bee/TundraBuildState.state.map
#	Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.dll
#	Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.pdb
#	Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.ref.dll
#	Library/Bee/backend1.traceevents
#	Library/Bee/backend2.traceevents
#	Library/Bee/buildprogram0.traceevents
#	Library/Bee/fullprofile.json
#	Library/Bee/tundra.digestcache
#	Library/Bee/tundra.log.json
#	Library/InspectorExpandedItems.asset
#	Library/SceneVisibilityState.asset
#	Library/ScriptAssemblies/Assembly-CSharp.dll
#	Library/ScriptAssemblies/Assembly-CSharp.pdb
#	Library/Search/transactions.db
#	Library/ShaderCache.db
#	Library/SourceAssetDB
#	Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
#	Library/UIElements/EditorWindows/UnityEditor.InspectorWindow.pref
#	Library/expandedItems
#	Library/ilpp.pid
#	Logs/AssetImportWorker0-prev.log
#	Logs/AssetImportWorker0.log
#	Logs/AssetImportWorker1-prev.log
#	Logs/AssetImportWorker1.log
#	Logs/AssetImportWorker2.log
#	Logs/AssetImportWorker3.log
#	Logs/shadercompiler-UnityShaderCompiler.exe0.log
#	Logs/shadercompiler-UnityShaderCompiler.exe1.log
#	Logs/shadercompiler-UnityShaderCompiler.exe2.log
#	UserSettings/Layouts/default-2022.dwlt
#	obj/Debug/Assembly-CSharp.csproj.AssemblyReference.cache

+++ Assets/02_Scripts/GameManager.cs
@@ -22,7 +22,7 @@ public class GameManager : MonoBehaviour
     [Header("## -- Player_Statas_List -- ##")]
     public float player_hp;
     public float player_max_hp;
-    public int bullet_damage;
+    public float bullet_damage;
     public int bullet_count;
     public float attack_delay;
     public int gold_count;
